requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
alembic==1.13.2
altair==5.4.0
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
appnope @ file:///home/conda/feedstock_root/build_artifacts/appnope_1707233003401/work
asgiref==3.8.1
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
attrs==24.2.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.34.158
botocore==1.34.158
build==1.2.1
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
cohere==5.8.0
coloredlogs==15.0.1
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
crewai==0.41.1
crewai-tools==0.4.26
ctransformers==0.2.27
dataclasses-json==0.5.14
debugpy @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_30hp2nowkm/croot/debugpy_1690905056188/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
docx2txt==0.8
embedchain==0.1.120
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
fastapi==0.112.0
fastavro==1.9.5
filelock==3.15.4
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.61.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-generativeai==0.7.2
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
gptcache==0.1.44
greenlet==3.0.3
grpc-google-iam-v1==0.13.1
grpcio==1.65.4
grpcio-status==1.62.3
grpcio-tools==1.62.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.24.5
humanfriendly==10.0
hyperframe==6.0.1
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
iniconfig==2.0.0
instructor==1.3.3
ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1719845458456/work
ipython @ file:///Users/runner/miniforge3/conda-bld/ipython_1680185563420/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.4
jiter==0.4.2
jmespath==1.0.1
joblib==1.4.2
json_repair==0.25.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1716472197302/work
jupyter_core @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_80l8bw91xp/croot/jupyter_core_1718818299834/work
kubernetes==30.1.0
lancedb==0.5.7
langchain==0.2.12
langchain-cohere==0.1.9
langchain-community==0.2.11
langchain-core==0.2.29
langchain-experimental==0.0.64
langchain-google-genai==1.0.8
langchain-huggingface==0.0.3
langchain-openai==0.1.21
langchain-pinecone==0.1.3
langchain-text-splitters==0.2.2
langchainplus-sdk==0.0.20
langsmith==0.1.98
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mdurl==0.1.2
# Editable install with no version control (Medical-Chatbot==0.0.0)
-e /Users/kumarrohit/Desktop/Langchain
mem0ai==0.0.9
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
narwhals==1.3.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.3
nltk==3.8.2
nodeenv==1.9.1
numexpr==2.10.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.40.3
openapi-schema-pydantic==1.2.4
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
orjson==3.10.7
outcome==1.3.0.post0
overrides==7.7.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1718189413536/work
pandas==2.2.2
parameterized==0.9.0
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==10.4.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.0.3
pinecone-plugin-interface==0.0.7
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1715777629804/work
pluggy==1.5.0
portalocker==2.10.1
posthog==3.5.0
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1718047967974/work
proto-plus==1.24.0
protobuf==4.25.4
psutil @ file:///Users/ec2-user/ci_py311_2/psutil_1679337242143/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pulsar-client==3.5.0
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1721585709575/work
py==1.11.0
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
pylance==0.9.18
pyparsing==3.1.2
pypdf==4.3.1
PyPika==0.48.9
pyproject_hooks==1.1.0
pyright==1.1.375
pysbd==0.3.4
PySocks==1.7.1
pytest==8.3.2
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
python-dotenv==1.0.1
pytube==15.0.0
pytz==2024.1
PyYAML==6.0.2
pyzmq @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_c7mhasdker/croot/pyzmq_1705605099593/work
qdrant-client==1.10.1
ratelimiter==1.2.0.post0
referencing==0.35.1
regex==2023.12.25
requests==2.32.3
requests-oauthlib==2.0.0
retry==0.9.2
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.4
schema==0.7.7
scikit-learn==1.5.1
scipy==1.14.0
selenium==4.23.1
semver==3.0.2
sentence-transformers==3.0.1
sentencepiece==0.2.0
shapely==2.0.5
shellingham==1.5.4
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.32
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
starlette==0.37.2
streamlit==1.37.1
streamlit-chat==0.1.1
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
torch==2.2.2
torchvision==0.17.2
tornado @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_7e0ruwlfee/croot/tornado_1718740118099/work
tqdm==4.66.5
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
transformers==4.44.0
trio==0.26.2
trio-websocket==0.11.1
typer==0.12.3
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.5
uvloop==0.19.0
watchfiles==0.23.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
websocket-client==1.8.0
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1718013267051/work
